backend/requirements.txt
Flask==2.3.3
pandas==2.3.1
python-dotenv==1.0.0
Flask-CORS==4.0.0
paho-mqtt==1.6.1
requests==2.31.0
cryptography==45.0.6
APScheduler==3.10.4
websocket-client==1.6.3
SQLAlchemy==1.4.46
Flask-Sock==0.7.0
simple-websocket==1.0.0
python-keycloak==3.7.0
PyJWT==2.8.0
requests-oauthlib==1.3.1
confluent-kafka==2.3.0
kafka-python==2.0.2
Flask-Limiter==3.5.0
